+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.23.1-0 / Unreleased
   * [Fixed] #695: List AMD dependency on jQuery UI
   * [Fixed] #735: Trying to set root node selected throws an error
+  * [Fixed] Drop marker for ext-glyph + ext-dnd5
 
 # 2.23.0 / 2017-05-27
   * **The external dependency on jQuery UI was removed**.